# core/openapi.yaml
openapi: 3.0.0
info:
  title: WACAST Core API
  description: WhatsApp Multi-Session Gateway API - Enterprise WhatsApp messaging platform
  version: 1.0.0
  contact:
    name: WACAST Support
    email: support@wacast.io
  license:
    name: Apache 2.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
  - url: https://api.wacast.io
    description: Production Server
...
paths:
...
  /api/v1/sessions/{device_id}:
    get:
      tags:
        - Sessions
      summary: Get session status
      description: Returns current status of a specific WhatsApp session
      operationId: getSessionStatus
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
          description: Device/Session identifier
          example: device001
      responses:
        "200":
          description: Session status retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionStatusResponse"
        "404":
          description: Session not found
  /api/v1/sessions/initiate:
    post:
      tags:
        - Sessions
      summary: Initiate new WhatsApp session
      description: |
        Start a new WhatsApp session for a device. Returns QR code for scanning.
        QR code is delivered via websocket or webhook.
      operationId: initiateSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiateSessionRequest"
            example:
              device_id: device001
              user_id: user123
              phone: "62812345678"
      responses:
        "200":
          description: Session initiated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InitiateSessionResponse"
        "400":
          description: Invalid request body
        "409":
          description: Session already exists for device
  /api/v1/sessions/{device_id}/stop:
    post:
      tags:
        - Sessions
      summary: Stop WhatsApp session
      description: Disconnect and remove an active WhatsApp session
      operationId: stopSession
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
          description: Device/Session identifier
      responses:
        "200":
          description: Session stopped successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Session stopped
                  device_id:
                    type: string
        "404":
          description: Session not found
...
components:
  schemas:
...
    SessionStatusResponse:
      type: object
      properties:
        device_id:
          type: string
        status:
          type: integer
        is_active:
          type: boolean
        last_activity:
          type: integer
    InitiateSessionRequest:
      type: object
      required:
        - device_id
        - user_id
        - phone
      properties:
        device_id:
          type: string
          description: Unique device identifier
        user_id:
          type: string
          description: User identifier
        phone:
          type: string
          description: Phone number (with country code, e.g., 62812345678)
    InitiateSessionResponse:
      type: object
      properties:
        message:
          type: string
          example: Session initiated
        device_id:
          type: string
        status:
          type: integer
          example: 2
...